refactor(architecture): collect core steps into core-workflow-steps
