+++ antenna-workflow-steps/pom.xml
@@ -27,9 +27,7 @@
         <module>analyzers/antenna-csv-analyzer</module>
         <module>analyzers/antenna-json-analyzer</module>
         <module>analyzers/antenna-ort-result-analyzer</module>
-        <module>analyzers/antenna-conf-analyzer</module>
         <!-- processors -->
-        <module>processors/antenna-conf-processor</module>
         <module>processors/abstract-antenna-compliance-checker</module>
         <module>processors/enricher/antenna-license-knowledgebase-resolver</module>
         <module>processors/enricher/antenna-license-resolver</module>
@@ -44,7 +42,5 @@
         <module>generators/antenna-csv-generator</module>
         <module>generators/antenna-HTML-report-generator</module>
         <module>generators/antenna-source-zip-generator</module>
-        <!-- outputHandlers -->
-        <module>outputHandlers/antenna-add-to-archive-output-handler</module>
     </modules>
 </project>
