+++ antenna/core-workflow-steps/src/main/java/org/eclipse/sw360/antenna/workflow/processors/checkers/ConfigurationChecker.java
