+++ antenna/core-workflow-steps/src/main/java/org/eclipse/sw360/antenna/workflow/processors/filter/ConfigurationHandlerRemove.java
