+++ antenna/core-workflow-steps/src/test/java/org/eclipse/sw360/antenna/analysis/filter/FilterTest.java
