+++ antenna/core-workflow-steps/src/test/java/org/eclipse/sw360/antenna/workflow/processors/filter/CommonConfigurationHandlerTest.java
