+++ pom.xml
@@ -47,6 +47,7 @@
         <module>antenna/model</module>
         <module>antenna/core</module>
         <module>antenna/frontend-stubs</module>
+        <module>antenna/core-workflow-steps</module>
         <module>modules</module>
         <module>antenna-testing</module>
         <module>antenna-license-provider</module>
